fix issues from review
- remove print statements
- remove unused imports
- remove dead code
- remove unused files
- fix return type for timelock from int64 to int
- export aes codec as part of package

+++ lib/src/schema/rad_request.dart
@@ -85,7 +85,7 @@ class RADRequest extends GeneratedMessage {
         "retrieve":
             List<dynamic>.from(retrieve.map((x) => x.jsonMap(asHex: asHex))),
         "tally": tally.jsonMap(asHex: asHex),
-        "time_lock": timeLock,
+        "time_lock": timeLock.toInt(),
       };
 
   Uint8List get pbBytes => writeToBuffer();
